pyproject.toml
[project]
name = "opnsense-ansible"
version = "0.1.0"
description = "Ansible collection for managing OPNsense firewall via REST API"
authors = [{name = "Justin Refi", email = "justinrefi@gmail.com"}]
requires-python = ">=3.12"
dependencies = [
    "opnsense-py>=0.1.0",
]

[tool.uv]
package = false

[tool.uv.sources]
opnsense-py = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "respx>=0.21.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "ansible-core>=2.17.0",
]

[tool.pytest.ini_options]
addopts = "-m 'not integration'"
markers = ["integration: requires live OPNsense instance"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
